backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
sqlmodel>=0.0.14
httpx>=0.25.0
openai>=1.3.0
python-dotenv>=1.0.0
pydantic-settings>=2.1.0
